+++ src/js/navigation.js
@@ -1,6 +1,5 @@
 /* global define */
 
-// TODO: Demo tabs
 // TODO: Focus handling
 
 (function(factory) {